+++ engine/flink/components-api/src/main/scala/pl/touk/nussknacker/engine/flink/api/process/FlinkContextInitializingFunction.scala
@@ -2,8 +2,8 @@ package pl.touk.nussknacker.engine.flink.api.process
 
 import org.apache.flink.api.common.functions.{OpenContext, RichMapFunction, RuntimeContext}
 import pl.touk.nussknacker.engine.api.Context
-import pl.touk.nussknacker.engine.api.process.{ContextInitializer, ContextInitializingFunction}
-import pl.touk.nussknacker.engine.api.runtimecontext.EngineRuntimeContext
+import pl.touk.nussknacker.engine.api.process.ContextInitializer
+import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
 
 class FlinkContextInitializingFunction[Raw](
     contextInitializer: ContextInitializer[Raw],
@@ -11,15 +11,15 @@ class FlinkContextInitializingFunction[Raw](
     convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
 ) extends RichMapFunction[Raw, Context] {
 
-  private var initializingStrategy: ContextInitializingFunction[Raw] = _
+  private var contextIdGenerator: ContextIdGenerator = _
 
   override def open(openContext: OpenContext): Unit = {
-    val contextIdGenerator = convertToEngineRuntimeContext(getRuntimeContext).contextIdGenerator(nodeId)
-    initializingStrategy = contextInitializer.initContext(contextIdGenerator)
+    contextIdGenerator = convertToEngineRuntimeContext(getRuntimeContext).contextIdGenerator(nodeId)
   }
 
   override def map(input: Raw): Context = {
-    initializingStrategy(input)
+    Context(contextIdGenerator.nextContextId())
+      .withVariables(contextInitializer.convertToInitialVariables(input).variables)
   }
 
 }