+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/EventGeneratorSourceFactory.scala
@@ -20,7 +20,6 @@ import pl.touk.nussknacker.engine.api.json.decoders.FromJsonTypingResultBasedDec
 import pl.touk.nussknacker.engine.api.json.encoders.ToJsonEncoder
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process._
-import pl.touk.nussknacker.engine.api.runtimecontext.ContextIdGenerator
 import pl.touk.nussknacker.engine.api.test.{TestData, TestRecord, TestRecordParser}
 import pl.touk.nussknacker.engine.api.typed.{typing, ReturningType}
 import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
@@ -136,10 +135,12 @@ class EventGeneratorSourceFactory(customTimestampAssigner: TimestampWatermarkHan
       }
 
       // This is a custom ContextInitializer, which initializes Context ignoring input.
-      // It is required, because in EventGenerator we fist initialize Context, and only then generate input.
+      // It is required, because in EventGenerator we first initialize Context, and only then generate input.
       private def contextInitializer[T]: ContextInitializer[T] = new ContextInitializer[T] {
-        override def initContext(contextIdGenerator: ContextIdGenerator): ContextInitializingFunction[T] =
-          (_: T) => Context(contextIdGenerator.nextContextId())
+        override def convertToInitialVariables(raw: T): ContextVariables = {
+          ContextVariables(Map.empty)
+        }
+
         override def validationContext(
             context: ValidationContext
         )(implicit nodeId: NodeId): ValidatedNel[ProcessCompilationError, ValidationContext] =